+++ app/Client/GuzzleAdUserClient.php
@@ -72,6 +72,7 @@ final class GuzzleAdUserClient implements AdUser
             return new UserContext(
                 $partialContext->keywords(),
                 AdUser::HUMAN_SCORE_ON_MISSING_TID,
+                AdUser::PAGE_RANK_ON_MISSING_TID,
                 Utils::hexUserId()
             );
         }
@@ -96,6 +97,7 @@ final class GuzzleAdUserClient implements AdUser
             return new UserContext(
                 $partialContext->keywords(),
                 AdUser::HUMAN_SCORE_ON_CONNECTION_ERROR,
+                AdUser::PAGE_RANK_ON_CONNECTION_ERROR,
                 Utils::hexUuidFromBase64UrlWithChecksum($partialContext->trackingId())
             );
         }