+++ CHANGELOG.md
@@ -7,8 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.3.2] - 2019-11-05
+## [1.3.2] - 2019-11-12
+### Added
+- Page rank handling
+- Processing payment reports selected by id
+### Changed
+- Export to adselect
+- Processing event payments on supply-side
+- Computing statistics of publishers
 ### Fixed
+- Conversion deletion during campaign change
+- Processing payment reports while events were not exported
 - Payment details for missing payments
 
 ## [1.3.1] - 2019-10-30